# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for HRD Attendance System",
        "title": "HRD Attendance API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/media/upload": {
            "post": {
                "description": "Upload image file and store metadata to database",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Upload image to imgbb",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Upload file (file/image/media)",
                        "name": "file",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Upload image",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Upload media",
                        "name": "media",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Media"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/overtime": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get overtime list with filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Overtime"
                ],
                "summary": "Get All Overtime",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/modelDto.AttendanceListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Employee request overtime",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Overtime"
                ],
                "summary": "Create Overtime Request",
                "parameters": [
                    {
                        "description": "Overtime Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateOvertimeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.OvertimeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/overtime/approve/{id}": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Manager/Admin approve overtime",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Overtime"
                ],
                "summary": "Approve Overtime Request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Overtime ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Notes",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ApproveOvertimeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.OvertimeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/overtime/reject/{id}": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Manager/Admin reject overtime",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Overtime"
                ],
                "summary": "Reject Overtime Request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Overtime ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Notes",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ApproveOvertimeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.OvertimeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/overtime/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Overtime"
                ],
                "summary": "Get Overtime By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.OvertimeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ApproveOvertimeRequest": {
            "type": "object",
            "properties": {
                "admin_notes": {
                    "type": "string"
                }
            }
        },
...
        "model.AttendanceResponse": {
            "type": "object",
            "properties": {
                "clock_in_latitude": {
                    "type": "number"
                },
                "clock_in_longitude": {
                    "type": "number"
                },
                "clock_in_media_url": {
                    "type": "string"
                },
                "clock_in_time": {
                    "type": "string"
                },
                "clock_out_latitude": {
                    "type": "number"
                },
                "clock_out_longitude": {
                    "type": "number"
                },
                "clock_out_media_url": {
                    "type": "string"
                },
                "clock_out_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.AttendanceStatus"
                },
                "user": {
                    "$ref": "#/definitions/model.UserResponse"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.AttendanceStatus": {
            "type": "string",
            "enum": [
                "working",
                "done",
                "late"
            ],
            "x-enum-varnames": [
                "StatusWorking",
                "StatusDone",
                "StatusLate"
            ]
        },
...
        "model.CreateOvertimeRequest": {
            "type": "object",
            "required": [
                "date",
                "end_time",
                "reason",
                "start_time"
            ],
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2026-04-08"
                },
                "end_time": {
                    "type": "string",
                    "example": "19:00"
                },
                "reason": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string",
                    "example": "17:00"
                }
            }
        },
...
        "model.ImageMeta": {
            "type": "object",
            "properties": {
                "extension": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "mime": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "model.Media": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imgbb_id": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/model.Meta"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "model.Meta": {
            "type": "object",
            "properties": {
                "delete_url": {
                    "type": "string"
                },
                "display_url": {
                    "type": "string"
                },
                "expiration": {
                    "type": "integer"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "$ref": "#/definitions/model.ImageMeta"
                },
                "size": {
                    "type": "integer"
                },
                "thumb": {
                    "$ref": "#/definitions/model.ImageMeta"
                },
                "time": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "url_viewer": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "model.OvertimeResponse": {
            "type": "object",
            "properties": {
                "admin_notes": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.OvertimeStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.UserResponse"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.OvertimeStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected"
            ],
            "x-enum-varnames": [
                "OvertimeStatusPending",
                "OvertimeStatusApproved",
                "OvertimeStatusRejected"
            ]
        },
...
        "model.RecentActivityResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "model.RoleResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.TenantResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tenant_settings": {
                    "$ref": "#/definitions/model.TenantSetting"
                }
            }
        },
        "model.TenantSetting": {
            "type": "object",
            "properties": {
                "allow_multiple_check": {
                    "type": "boolean",
                    "example": false
                },
                "allow_remote": {
                    "type": "boolean",
                    "example": false
                },
                "clock_in_end_time": {
                    "type": "string",
                    "example": "09:00"
                },
                "clock_in_start_time": {
                    "type": "string",
                    "example": "07:00"
                },
                "clock_out_end_time": {
                    "type": "string",
                    "example": "23:00"
                },
                "clock_out_start_time": {
                    "type": "string",
                    "example": "16:00"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-07T13:21:24Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "late_after_minute": {
                    "type": "integer",
                    "example": 480
                },
                "max_radius_meter": {
                    "type": "number",
                    "example": 100
                },
                "office_latitude": {
                    "type": "number",
                    "example": -6.1339179
                },
                "office_longitude": {
                    "type": "number",
                    "example": 106.8329504
                },
                "require_location": {
                    "type": "boolean",
                    "example": true
                },
                "require_selfie": {
                    "type": "boolean",
                    "example": true
                },
                "tenant": {
                    "$ref": "#/definitions/model.Tenant"
                },
                "tenant_id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-04-07T13:21:24Z"
                }
            }
        },
...
        "model.UserResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Jl. Sudirman No. 1"
                },
                "attendances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AttendanceResponse"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "department": {
                    "type": "string",
                    "example": "IT"
                },
                "email": {
                    "type": "string",
                    "example": "budi@company.com"
                },
                "employee_id": {
                    "type": "string",
                    "example": "FT-001"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "media_url": {
                    "type": "string",
                    "example": "https://cdn.example.com/profile/budi.jpg"
                },
                "name": {
                    "type": "string",
                    "example": "Budi Santoso"
                },
                "phone_number": {
                    "type": "string",
                    "example": "08123456789"
                },
                "recent_activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RecentActivityResponse"
                    }
                },
                "role": {
                    "$ref": "#/definitions/model.RoleResponse"
                },
                "tenant": {
                    "$ref": "#/definitions/model.TenantResponse"
                },
                "tenant_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "modelDto.AttendanceListResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/modelDto.Meta"
                }
            }
        },
        "modelDto.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "modelDto.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}